Web_site_'Babam'/requirements.txt
bcrypt==4.2.0
Flask==2.1.2
mysql-connector-python==9.0.0
Werkzeug==2.1.2